Moving user agent calculation into bigtable-client-core. We always want to include cloud bigtable client version and jdk version in the user agent.  The BigtableOptions user agent is an additional piece of information rather than a substitute, so treat it that way.
