+++ bigtable-client-core-parent/bigtable-client-core/src/main/resources/com/google/cloud/bigtable/config/bigtable-version.properties
@@ -1,0 +1,2 @@
+bigtable.version=${project.version}
+
