+++ bigtable-hbase-parent/bigtable-hbase-1.1/pom.xml
@@ -64,13 +64,6 @@ limitations under the License.
         </dependency>
     </dependencies>
     <build>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <!-- enable project.version substitution. -->
-                <filtering>true</filtering>
-            </resource>
-        </resources>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
