--- bigtable-hbase-parent/bigtable-hbase-1.2/src/main/resources/com/google/cloud/bigtable/hbase/bigtable-hbase.properties
@@ -1,1 +1,0 @@
-bigtable.hbase.user_agent=${project.artifactId}-${project.version}
