+++ bigtable-hbase-parent/bigtable-hbase-mapreduce/pom.xml
@@ -82,13 +82,6 @@ limitations under the License.
         </dependency>
     </dependencies>
     <build>
-        <resources>
-            <resource>
-                <directory>src/main/resources</directory>
-                <!-- enable project.version substitution. -->
-                <filtering>true</filtering>
-            </resource>
-        </resources>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
